Alireza Changed Price DataType To Bigint And Also Changed Product Deatils View

+++ AlirezaEShop/Models/AddEditViewModel.cs
@@ -10,7 +10,7 @@ namespace AlirezaEShop.Models
         public int ID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public Int64 Price { get; set; }
         public int QuantityInStock { get; set; }
         public IFormFile? Picture { get; set; }
         public string? PictureExtention { get; set; }
